+++ src/components/mypage/MyTrade.tsx
@@ -1,13 +1,13 @@
-import data, { TradeData } from "../../data/TradeData";
-import { ReactComponent as PriceIcon } from "../../assets/icon/price.svg";
-import { ReactComponent as SalerIcon } from "../../assets/icon/saler.svg";
-import { ReactComponent as ChatIcon } from "../../assets/icon/chat-color.svg";
-import "../../styles/mypage/MyTrade.scss";
+import data, { TradeData } from "@/data/TradeData";
+import PriceIcon from "@/assets/icon/price.svg?react";
+import SalerIcon from "@/assets/icon/saler.svg?react";
+import ChatIcon from "@/assets/icon/chat-color.svg?react";
+import "@/styles/mypage/MyTrade.scss";
 import { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
-import { fetchNoMajorTokenData } from "../../api/fetchData";
-import { BookData } from "../../types/Types";
-import { postSold } from "../../api/postData";
+import { fetchNoMajorTokenData } from "@/api/fetchData";
+import { BookData } from "@/types/Types";
+import { postSold } from "@/api/postData";
 
 const MyTrade = () => {
   const [tradeData, setTradeData] = useState<BookData[]>([]);
