+++ src/components/study/PopularStudy.tsx
@@ -1,8 +1,8 @@
-import "../../styles/study/PopularStudy.scss";
+import "@/styles/study/PopularStudy.scss";
 import { useNavigate } from "react-router-dom";
 import { useState, useEffect } from "react";
-import { fetchData } from "../../api/fetchData";
-import { StudyData } from "../../types/Types";
+import { fetchData } from "@/api/fetchData";
+import { StudyData } from "@/types/Types";
 
 const PopularStudy = () => {
   const [studyData, setStudyData] = useState<StudyData[]>([]);
@@ -28,7 +28,7 @@ const PopularStudy = () => {
           <li key={index} onClick={() => handleItemClick(item.id)}>
             <p>{item.title}</p>
             <div>
-              <img src={require("../../assets/img/dku-logo.png")} alt="logo" />
+              <img src={require("@/assets/img/dku-logo.png")} alt="logo" />
               <p>
                 {item.school_name} {item.major_name}{" "}
                 {String(item.admission_date).slice(-2)}학번
